third_party/blink/web_tests/http/tests/misc/resource-timing-iframe-restored-from-history.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<script>
description('Tests that iframe restored from history does not report resource timing.');
window.jsTestIsAsync = true;

function runTest() {
    if (!sessionStorage.didNav) {
        sessionStorage.didNav = true;
        // Navigate a timeout to make sure we generate a history entry that we can go back to.
        setTimeout(function() {
            location.href = 'resources/alert-then-back.html';
        }, 0);
    } else {
        delete sessionStorage.didNav;
        resources = performance.getEntriesByType('resource');
        shouldBe('resources.length', '1');
        shouldBeEqualToString('resources[0].name', 'http://127.0.0.1:8000/js-test-resources/js-test.js');
        if (window.testRunner)
            finishJSTest();
    }
}

if (!sessionStorage.didNav) {
  window.onload = function() {
    document.querySelector("iframe").src = "resources/frame-initial-url.html";
  }
}

</script>
<iframe></iframe>
